requirements.txt
ipykernel
langchain
langchain-openai
langchain-community
langchain-core[tracing]
python-dotenv
pypdf
faiss-cpu
structlog
PyMuPDF
pyyaml
pandas
numpy
pytest
streamlit
-e .